Model_&_Deployment_Files/api/requirements.txt
flask
torch
torchvision
pillow
google-cloud-storage